OthelloView
